feat: native build mode; speed up build process
The tl;dr for the native build mode is it allows one to build baml-rest
while already in the context of a Docker build by starting from the
builder base image (not published yet) that contains a prebuilt version
of `baml-rest` and all necessary dependencies.

To make this reasonably maintainable, the previously multi-step
Dockerfile build process is now a single Bash script, and caching is
fully handled via a cache mount, which appears to be performing _much_
better than the speed we had by relying on layer caching (not that
the Dockerfile was ever super well optimized, but there's that, it's
a win-win).