fix: unwrap baml dynamic classes/enums/unions
When using the `type_builder` to define, e.g., a field that is a custom
union, instead of getting the value under `DynamicProperties`, you get
a struct with `Variant` and `Value` properties, which isn't particularly
nice to consume and is also an internal BAML runtime detail that the
consumer of our API shouldn't need to care about.

It just so turns out that these dynamic values are present under the
`serde` subpackage of the BAML runtime, so detecting them is easy. This
patch adds a new `UnwrapDynamicValue` helper that is applied to dynamic
values coming out of the BAML runtime. It is a recursive helper that
ensures all values that reach the end user are normalized.